+++ xmake/core/ui/curses.lua
@@ -213,5 +213,10 @@ function curses.cursor_set(state)
     end
 end
 
+-- has mouse?
+function curses.has_mouse()
+    return curses.KEY_MOUSE and true or false
+end
+
 -- return module: curses
 return curses